Move ThemeConfigurationWidget
