+++ portal/src/ThemeConfigurationWidget.module.scss
